Moved thirdparty out of v1 and delete leftovers
